+++ custom_components/zbrano/const.py
@@ -1,5 +1,7 @@
 DOMAIN = "zbrano"
 CONF_URL = "url"
 CONF_TOKEN = "token"
+CONF_FALLBACK_AGENT = "fallback_agent"
 DEFAULT_URL = "http://homeassistant.local:8099"
 REQUEST_TIMEOUT = 180
+HEALTH_TIMEOUT = 3